Updated the board layout. Changed the options dialog UI. Added options to change board dimensions and name the board. Implemented SQL database to save boards.

+++ app/src/main/java/com/akashd50/lb/objects/Camera.java
@@ -14,7 +14,7 @@ public class Camera {
 
     private SimpleVector distanceFromOrigin;
     private boolean simulateCCUMovement;
-    private int width, height;
+    private float width, height;
     public Camera(){
         cameraRotation = new SimpleVector(0f,0f,0f);
         distanceFromOrigin = new SimpleVector(0f,0f,5f);
@@ -31,7 +31,7 @@ public class Camera {
         mMVPMatrix = mvp;
     }
 
-    public void setAdditionalParms(int w, int h){
+    public void setAdditionalParms(float w, float h){
         width = w;
         height = h;
         float ratio = (float)height/(float)width;
